timestamp es long long, lissandra usa hilos

+++ Kernel/paquete.h
@@ -50,7 +50,7 @@ typedef struct t_paquete_insert
 	uint16_t valor_key;
 	uint32_t value_long; 		//Longitud del value
 	char* value;
-	int timestamp;
+	long long timestamp;
 }__attribute__((packed)) t_paquete_insert;
 
 //---------------------ARCHIVO LOGGER/CONFIG
@@ -67,7 +67,7 @@ int codigo_ingresado(char* lineaRequest);
 
 //---------------------CREAR PAQUETE
 t_paquete_select* crear_paquete_select(char *nombretabla,uint16_t valor_key);
-t_paquete_insert* crear_paquete_insert(char *nombretabla,uint16_t valor_key, char *value, int timestamp);
+t_paquete_insert* crear_paquete_insert(char *nombretabla,uint16_t valor_key, char *value, long long timestamp);
 
 //---------------------ARMAR PAQUETE
 t_paquete_select* selectf(char* parametros);
